pyproject.toml
[project]
name = "rag-homework"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.6",
    "langchain-community>=0.3.30",
    "langchain-ollama>=0.3.10",
    "langchain-openai>=0.3.34",
    "pypdf>=6.1.1",
    "python-dotenv>=1.1.1",
]
